+++ includes/scripts/_on_build_cleanup.sh
@@ -1,0 +1,19 @@
+#!/bin/sh
+
+# Remove apt lists
+rm -rf /var/lib/apt/lists/*
+
+# Locale cleanup
+cp -R /usr/share/locale/en\@* /tmp/ && rm -rf /usr/share/locale/* && mv /tmp/en\@* /usr/share/locale/
+
+# Clean out docs
+rm -rf /usr/share/doc /usr/share/doc-base /usr/share/man /usr/share/locale /usr/share/zoneinfo /var/cache/debconf/*-old
+
+# Clean out package management dirs
+rm -rf /var/lib/cache /var/lib/log
+
+# Clean out /tmp
+rm -rf /tmp/*
+
+# Clear npm cache
+npm cache clear