+++ includes/scripts/_setup_tupperbuild.sh
@@ -2,7 +2,7 @@
 
 cd $TUPPERBUILD_DIR
 
-echo "> Setting up tupperbuild..."
+echo "[-] Setting up tupperbuild..."
 
 npm install
 check_code $?