+++ includes/scripts/start_app.sh
@@ -5,7 +5,7 @@ BASEDIR=`dirname $0`
 . $BASEDIR/_common.sh
 
 if [ ! -f $OUTPUT_DIR/bundle/main.js ]; then
-  echo "There is no bundle. Please see usage docs here: https://github.com/chriswessels/meteor-tupperware"
+  echo "[!] There is no application bundle. Please see usage docs here: https://github.com/chriswessels/meteor-tupperware"
   exit 1
 fi
 
@@ -19,6 +19,6 @@ if [ -z "$METEOR_ENV" ]; then
   export METEOR_ENV="production"
 fi
 
-echo "> meteor-tupperware is starting your application with NODE_ENV=$NODE_ENV and METEOR_ENV=$METEOR_ENV..."
+echo "[-] meteor-tupperware is starting your application with NODE_ENV=$NODE_ENV and METEOR_ENV=$METEOR_ENV..."
 
 exec node $OUTPUT_DIR/bundle/main.js "$@"